+++ src/main/java/lotto/LottoApplication.java
@@ -24,7 +24,7 @@ public class LottoApplication {
         List<LottoTicket> lottoTickets = generator.generateLottoTickets(ticketCount);
         resultView.printLottoTickets(lottoTickets);
 
-        Scanner scanner = new Scanner(inputView.inputWinningNumbers(), inputView.inputBonusNumber());
+        Scanner.newScanner(inputView.inputWinningNumbers(), inputView.inputBonusNumber());
         LottoResult lottoResult = LottoResult.getLottoResult(lottoTickets);
         resultView.printLottoResult(lottoResult);
     }
